Moves RequestType to Fix Circular Dependency
