+++ src/lib/bridge/handlers/execute_script.ts
@@ -1,5 +1,5 @@
 import {EmptyResponseHandler, PrivilegedHandler} from './main';
-import {RequestType} from './handlers';
+import {RequestType} from './types';
 
 interface ExecuteScriptRequest {
     path: string;
