+++ src/lib/bridge/handlers/fetch_inspect_info.ts
@@ -1,5 +1,5 @@
 import {SimpleHandler} from './main';
-import {RequestType} from './handlers';
+import {RequestType} from './types';
 
 interface Sticker {
     slot: number;
