+++ src/lib/bridge/handlers/fetch_skin_model.ts
@@ -1,5 +1,5 @@
 import {SimpleHandler} from './main';
-import {RequestType} from './handlers';
+import {RequestType} from './types';
 
 export interface FetchSkinModelRequest {
     inspectLink: string;
