+++ src/lib/bridge/handlers/handlers.ts
@@ -9,19 +9,7 @@ import {RequestHandler} from '../types';
 import {FetchPendingTrades} from './fetch_pending_trades';
 import {FetchSkinModel} from './fetch_skin_model';
 import {StorageRemove} from './storage_remove';
-
-export enum RequestType {
-    EXECUTE_SCRIPT_ON_PAGE,
-    EXECUTE_CSS_ON_PAGE,
-    FETCH_INSPECT_INFO,
-    FETCH_STALL,
-    STORAGE_GET,
-    STORAGE_SET,
-    STORAGE_REMOVE,
-    CSMONEY_PRICE,
-    FETCH_PENDING_TRADES,
-    FETCH_SKIN_MODEL,
-}
+import {RequestType} from './types';
 
 export const HANDLERS_MAP: {[key in RequestType]: RequestHandler<any, any>} = {
     [RequestType.EXECUTE_SCRIPT_ON_PAGE]: ExecuteScriptOnPage,
