+++ src/lib/bridge/handlers/types.ts
@@ -1,0 +1,12 @@
+export enum RequestType {
+    EXECUTE_SCRIPT_ON_PAGE,
+    EXECUTE_CSS_ON_PAGE,
+    FETCH_INSPECT_INFO,
+    FETCH_STALL,
+    STORAGE_GET,
+    STORAGE_SET,
+    STORAGE_REMOVE,
+    CSMONEY_PRICE,
+    FETCH_PENDING_TRADES,
+    FETCH_SKIN_MODEL,
+}
