+++ src/lib/bridge/server.ts
@@ -1,6 +1,7 @@
 import {InternalRequestBundle, RequestHandler, Version} from './types';
 import MessageSender = chrome.runtime.MessageSender;
-import {HANDLERS_MAP, RequestType} from './handlers/handlers';
+import {HANDLERS_MAP} from './handlers/handlers';
+import {RequestType} from './handlers/types';
 
 function findHandler(type: RequestType): RequestHandler<any, any> | undefined {
     return HANDLERS_MAP[type];
